backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
certifi==2026.2.25
click==8.3.1
Django==5.2.12
django-cors-headers==4.9.0
djangorestframework==3.16.1
exceptiongroup==1.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
nodeenv==1.10.0
prisma==0.15.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
sqlparse==0.5.5
StrEnum==0.4.15
tomlkit==0.14.0
typing-inspection==0.4.2
typing_extensions==4.15.0
